update6

+++ src/PortfolioApp.Domain/Options/JwtOption.cs
@@ -6,6 +6,6 @@ public sealed class JwtTokenOption
     public string SecretKey { get; set; } = string.Empty;
     public string Issuer { get; set; } = string.Empty;
     public string Audience { get; set; } = string.Empty;
-    public double AccessTokenExpirationMinutes { get; set; }
+    public int AccessTokenExpirationMinutes { get; set; }
     public int RefreshTokenExpirationDays { get; set; }
 }
